transaction posts- api conversions

+++ core/slot.go
@@ -4,10 +4,10 @@ import "time"
 
 //GetTime return time slot difference in secods. This timestamp is
 //added to the transaction.
-func GetTime() float64 {
+func GetTime() int32 {
 	mainNetStart := time.Date(2017, 3, 21, 13, 00, 0, 0, time.UTC)
 	now := time.Now()
 
 	diff := now.Sub(mainNetStart)
-	return diff.Seconds()
+	return int32(diff.Seconds())
 }
